refactor: simplify CommandPalette integration with custom hook
- Replace confusing render prop pattern with simple prop interface
- Create useCommandPalette hook to manage keyboard shortcut and state
- Move keyboard event listener to custom hook for better separation of concerns
- Make _app.tsx much more readable and straightforward
- CommandPalette now accepts simple open/setOpen props instead of children function
