+++ src/pages/_app.tsx
@@ -2,6 +2,7 @@ import '@/styles/globals.css';
 import type { AppProps } from 'next/app';
 import { Geist } from 'next/font/google';
 import { CommandPalette } from '@/components/command/CommandPalette';
+import { useCommandPalette } from '@/hooks/useCommandPalette';
 
 const geistSans = Geist({
   variable: '--font-geist-sans',
@@ -9,9 +10,12 @@ const geistSans = Geist({
 });
 
 export default function App({ Component, pageProps }: AppProps) {
+  const { open, setOpen, openCommand } = useCommandPalette();
+
   return (
     <div className={geistSans.variable}>
-      <CommandPalette>{(setOpen) => <Component {...pageProps} onOpenCommand={() => setOpen(true)} />}</CommandPalette>
+      <Component {...pageProps} onOpenCommand={openCommand} />
+      <CommandPalette open={open} setOpen={setOpen} />
     </div>
   );
 }
